// client/employee-management/src/app/register/create-account-contact-person/reference.ts
export class Reference{
    re_firstname: string = '';
    re_lastname: string = '';
    re_middlename: string = '';
    re_cellphone: string = '';
    re_email: string = '';
    re_address1: string = '';
    re_address2: string = '';
    re_city: string = '';
    re_zipcode: string = '';
    re_state: string = '';
    re_relationship: string = '';
}